[236-implement-redisset-data-structure-as-a-special-field-type] - rename file
